pyproject.toml
[tool.poetry]
name = "axiom"
version = "0.1.0"
description = "Axiom AI consulting intelligence system"
authors = ["Connors-EO"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
psycopg2-binary = "^2.9"
python-dotenv = "^1.0"
boto3 = "^1.34"
httpx = "^0.27"

[tool.poetry.scripts]
db-migrate = "backend.db.migrate:main"
db-seed = "backend.db.seed:main"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
pytest-cov = "^5.0"
pytest-asyncio = "^0.23"
mypy = "^1.9"
ruff = "^0.4"
types-psycopg2 = "^2.9"
boto3-stubs = {extras = ["bedrock-runtime"], version = "^1.42.63"}

[tool.pytest.ini_options]
testpaths = ["backend/tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--cov=backend/src --cov-report=term-missing --cov-fail-under=100 --ignore=backend/tests/integration"
markers = [
    "integration: marks tests as integration tests (real AWS + DB)",
]

[tool.coverage.run]
source = ["backend/src"]
omit = ["*/tests/*"]

[tool.coverage.report]
fail_under = 100
show_missing = true

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = false

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
